fix(auth): dedupe activation HMAC helpers and prettier PR-04 leftovers
Also fix MD031 fences and prettier on R2 helpers that failed Lint after PR-13.
